Refactoring (main) - 43
