+++ controllers/contacts/index.js
@@ -2,6 +2,7 @@ const getAllContacts = require("./getAllContacts")
 const getContactById = require("./getContactById")
 const addContact = require("./addContact")
 const updatePutContact = require("./updatePutContact")
+const updatePatchContact = require("./updatePatchContact")
 
 
 
@@ -9,5 +10,6 @@ module.exports = {
     getAllContacts,
     getContactById,
     addContact,
-    updatePutContact
+    updatePutContact,
+    updatePatchContact
 }
