+++ models/contacts.json
@@ -1,86 +1,1 @@
-[
-  {
-    "id": "1",
-    "name": "Allen Raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792"
-  },
-  {
-    "id": "2",
-    "name": "Chaim Lewis",
-    "email": "dui.in@egetlacus.ca",
-    "phone": "(294) 840-6685"
-  },
-  {
-    "id": "3",
-    "name": "Kennedy Lane",
-    "email": "mattis.Cras@nonenimMauris.net",
-    "phone": "(542) 451-7038"
-  },
-  {
-    "id": "4",
-    "name": "Wylie Pope",
-    "email": "est@utquamvel.net",
-    "phone": "(692) 802-2949"
-  },
-  {
-    "id": "5",
-    "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
-    "phone": "(501) 472-5218"
-  },
-  {
-    "id": "6",
-    "name": "Abbot Franks",
-    "email": "scelerisque@magnis.org",
-    "phone": "(186) 568-3720"
-  },
-  {
-    "id": "7",
-    "name": "Reuben Henry",
-    "email": "pharetra.ut@dictum.co.uk",
-    "phone": "(715) 598-5792"
-  },
-  {
-    "id": "8",
-    "name": "Simon Morton",
-    "email": "dui.Fusce.diam@Donec.com",
-    "phone": "(233) 738-2360"
-  },
-  {
-    "id": "9",
-    "name": "Thomas Lucas",
-    "email": "nec@Nulla.com",
-    "phone": "(704) 398-7993"
-  },
-  {
-    "id": "10",
-    "name": "Alec Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
-  },
-  {
-    "id": "790c7900471f",
-    "name": "Alec_WWW Howar_RRR",
-    "email": "Donec.elementum_UUU@scelerisque.ua",
-    "phone": "(999) 206-2688"
-  },
-  {
-    "id": "5f67228eec6e",
-    "name": "Alec12 Howar_d12",
-    "email": "Donec.elementum_12@scelerisque.net",
-    "phone": "(222) 206-2688"
-  },
-  {
-    "id": "a0df5f5ce3f4",
-    "name": "Alec_333 Howar_333",
-    "email": "Donec.elementum__111@scelerisque.com",
-    "phone": "(111) 206-2688"
-  },
-  {
-    "id": "329a53a2ba73",
-    "name": "ReubenYYY HenryYYY",
-    "email": "pharetra.ut_YYY@dictum.co.ua",
-    "phone": "(123) 456-9999"
-  }
-]
+[{"id":"1","name":"Allen Raymond","email":"nulla.ante@vestibul.co.uk","phone":"(992) 914-3792"},{"id":"2","name":"Chaim Lewis","email":"dui.in@egetlacus.ca","phone":"(294) 840-6685"},{"id":"3","name":"Kennedy Lane","email":"mattis.Cras@nonenimMauris.net","phone":"(542) 451-7038"},{"id":"4","name":"Wylie Pope","email":"est@utquamvel.net","phone":"(692) 802-2949"},{"id":"5","name":"Cyrus Jackson","email":"nibh@semsempererat.com","phone":"(501) 472-5218"},{"id":"6","name":"Abbot Franks","email":"scelerisque@magnis.org","phone":"(186) 568-3720"},{"id":"7","name":"Reuben Henry","email":"pharetra.ut@dictum.co.uk","phone":"(715) 598-5792"},{"id":"8","name":"Simon Morton","email":"dui.Fusce.diam@Donec.com","phone":"(233) 738-2360"},{"id":"9","name":"Thomas Lucas","email":"nec@Nulla.com","phone":"(704) 398-7993"},{"id":"10","name":"Alec Howard","email":"Donec.elementum@scelerisquescelerisquedui.net","phone":"(748) 206-2688"},{"id":"790c7900471f","name":"Alec_WWW Howar_RRR","email":"Donec.elementum_UUU@scelerisque.ua","phone":"(999) 206-2688"},{"id":"5f67228eec6e","name":"Alec12 Howar_d12","email":"Donec.elementum_12@scelerisque.net","phone":"(222) 206-2688"},{"id":"a0df5f5ce3f4","name":"Alec_333 Howar_333","email":"Donec.elementum__111@scelerisque.com","phone":"(111) 206-2688"},{"id":"329a53a2ba73","name":"ReubenYYY HenryYYY","email":"pharetra.ut_QQQQ@dictum.co.ua","phone":"(123) 456-9999"}]
