+++ routes/api/contacts.js
@@ -4,51 +4,51 @@ const router = express.Router()
 const { NotFound } = require('http-errors')
 const Joi = require('joi')
 
-const contactsOperations = require("../../models/contacts")
+// const contactsOperations = require("../../models/contacts")
 
 const { contacts: ctrl } = require("../../controllers")
 
-const { lineBreak } = require("../../service");
+// const { lineBreak } = require("../../service");
 
 
 //-----------------------------------------------------------------------------
 //* ++++++++++++++++++++++ Схемы ВАЛИДАЦИИ Joi +++++++++++++++++++++++++
-const contactSchemaPostPut = Joi.object({
-  name: Joi.string()
-    // .alphanum()
-    .min(3)
-    .max(30)
-    .required(),
+// const contactSchemaPostPut = Joi.object({
+//   name: Joi.string()
+//     // .alphanum()
+//     .min(3)
+//     .max(30)
+//     .required(),
 
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
-    .required(),
+//   email: Joi.string()
+//     .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
+//     .required(),
 
-  phone: Joi.string()
-    // .alphanum()
-    .min(5)
-    .max(14)
-    .required(),
-});
+//   phone: Joi.string()
+//     // .alphanum()
+//     .min(5)
+//     .max(14)
+//     .required(),
+// });
 
 //--------------------------------------------------------------------
-const contactSchemaPutch = Joi.object({
-  name: Joi.string()
-    // .alphanum()
-    .min(3)
-    .max(30)
-    .optional(),
+// const contactSchemaPutch = Joi.object({
+//   name: Joi.string()
+//     // .alphanum()
+//     .min(3)
+//     .max(30)
+//     .optional(),
 
-  email: Joi.string()
-    .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
-    .optional(),
+//   email: Joi.string()
+//     .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
+//     .optional(),
 
-  phone: Joi.string()
-    // .alphanum()
-    .min(5)
-    .max(14)
-    .optional(),
-});
+//   phone: Joi.string()
+//     // .alphanum()
+//     .min(5)
+//     .max(14)
+//     .optional(),
+// });
 //* _______________________ Схемы ВАЛИДАЦИИ Joi _______________________
 
 
@@ -79,55 +79,7 @@ router.put('/:contactId', ctrl.updatePutContact);
 
 //-----------------------------------------------------------------------------
 //! 4-2. PATCH-Обновление ОДНОГО КОНТАКТА по id
-router.patch("/:contactId", async (req, res, next) => {
-  try {
-    //! ===========================console============================
-    console.log("START-->PATCH/:id".rainbow); //!
-    lineBreak();
-    //! ==============================================================
-
-    //* +++++++++++++++++++++++ ВАЛИДАЦИЯ Joi +++++++++++++++++++++++++++++
-    const validationResult = contactSchemaPutch.validate(req.body);
-
-    if (validationResult.error) {
-      //! ===========================console============================
-      console.log("Ошибка ВАЛИДАЦИИ:".bgRed.black);
-      console.log("");
-      console.log(validationResult.error);
-      lineBreak();
-      console.log("END-->PATCH/:id".rainbow); //!
-      //! ==============================================================
-      //! 1 - вариант
-      // return res.status(400).json({ "message": "missing required name field" });
-      //! 2 - вариант
-      // validationResult.error.status = 400
-      // throw validationResult.error
-      //! 3 - вариант
-      return res.status(400).json({ status: validationResult.error.details });
-    }
-    //* __________________________ ВАЛИДАЦИЯ Joi __________________________
-
-    const { contactId } = req.params;
-    const contact = await contactsOperations.updatePatchContact(contactId, req.body)
-
-    if (!contact) {
-      //! 4 - вариант
-      throw new NotFound(`Contact wiht id:'${contactId}' not found`)
-    }
-
-    res.status(200).json({
-      status: "success",
-      code: 200,
-      data: {
-        result: contact
-      }
-    })
-
-  } catch (e) {
-    next(e);
-    // res.status(500).json({ error: e.message });
-  }
-});
+router.patch("/:contactId", ctrl.updatePatchContact);
